comcpugpumhfemass

+++ CompareCPUGPU/compileCode.sh
@@ -1,11 +1,17 @@
 #/bin/bash
 
+if [[ -z "$1" ]]
+then
+    echo 'Please specify which cu file you want to compile (you can input 1, 2, 3 or 4)'
+    exit 0
+fi
+
 cd build
 # rm -rf CMakeCache.txt  CMakeFiles  cmake_install.cmake  Makefile
 cd ../bin 
 rm -rf main
 cd ..
 cd build 
-cmake ..
+cmake .. -DCUFILE=$1
 make
 cd ..
